docs: typo fixes Update getting-started.md (#348)
A couple of typographical errors in the documentation that could cause
confusion or make the text less professional.

1. In the section `::: info Not ready to build?`, the word
"Specifcation" was misspelled. The correct spelling is
**Specification**.
   
2. In the section `Complete the prompts and follow the instructions`,
the word "depedencies" was incorrectly spelled. The correct spelling is
**dependencies**.


<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on correcting minor typographical errors in the
documentation for getting started with `Frames`.

### Detailed summary
- Changed "Frames Specifcation" to "Frames Specification" in
`docs/developers/frames/getting-started.md`.
- Corrected "depedencies" to "dependencies" in the installation command
section.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->