Made initial changes (#414)
<!-- start pr-codex -->

## PR-Codex overview
This PR focuses on rebranding "Frames" to "Mini Apps" across the
documentation and adding a redirect for users. It clarifies the
transition and encourages users to adopt the new terminology and
resources.

### Detailed summary
- Updated text to refer to "Mini Apps" instead of "Frames" in
`docs/developers/index.md` and `docs/index.md`.
- Added a new section in `docs/.vitepress/config.mts` for a redirect
link to the `frames-redirect.md`.
- Created `docs/reference/frames-redirect.md` to inform users about the
rebranding and deprecation of Frames v2.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->