feat: add Chinese language support and translate content
This commit introduces Chinese language support to the documentation site by adding a new locale configuration in the VitePress config file.
It also includes the translation of all documentation pages into Chinese

The changes include:
- Added Chinese locale configuration in `.vitepress/config.mts`
- Translated all documentation pages into Chinese
- Added Chinese navigation and sidebar configurations

This enhancement aims to make the Farcaster documentation more accessible to Chinese-speaking developers and users.