fix: Link to FIPs github discussions instead of duplicating the list (#446)
FIPs list was getting out of date. Just link to the discussion board for
latest view

<!-- start pr-codex -->

---

## PR-Codex overview
This PR updates the documentation in `fips.md` to improve clarity and
accuracy regarding finalized FIPs. It modifies the text to better direct
readers to relevant links and removes a table that listed FIPs.

### Detailed summary
- Removed the table listing finalized FIPs and their details.
- Updated text to clarify where to find finalized FIPs.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your
question}`

<!-- end pr-codex -->